Backend/requirements.txt
django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
pillow==10.0.0
geopy==2.4.1